requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
appdirs==1.4.4
apturl==0.5.2
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.4.0
backoff==2.2.1
bcrypt==5.0.0
beautifulsoup4==4.14.2
beniget==0.4.1
blinker==1.4
Brlapi==0.8.3
Brotli==1.0.9
build==1.3.0
cachetools==6.2.1
certifi==2020.6.20
cffi==2.0.0
cfgv==3.4.0
chardet==4.0.0
charset-normalizer==3.4.4
chromadb==1.1.1
click==8.3.0
colorama==0.4.4
coloredlogs==15.0.1
command-not-found==0.3
contourpy==1.1.1
crewai==1.3.0
crewai-tools==1.3.0
cryptography==46.0.3
cupshelpers==1.0
cycler==0.12.0
dbus-python==1.2.18
decorator==4.4.2
defer==1.0.6
defusedxml==0.7.1
deprecation==2.1.0
diskcache==5.6.3
distlib==0.3.5
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==5.0.6
docker==7.1.0
docstring_parser==0.17.0
duplicity==0.8.21
durationpy==0.10
et_xmlfile==2.0.0
exceptiongroup==1.1.3
fasteners==0.14.1
filelock==3.7.1
flatbuffers==25.9.23
fonttools==4.43.0
frozenlist==1.8.0
fs==2.4.12
fsspec==2025.10.0
future==0.18.2
gast==0.5.2
google-auth==2.43.0
googleapis-common-protos==1.72.0
gpg==1.16.0
grpcio==1.76.0
h11==0.16.0
hf-xet==1.2.0
html5lib==1.1
httpcore==1.0.9
httplib2==0.20.2
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.1.2
humanfriendly==10.0
identify==2.6.15
idna==3.3
importlib_metadata==8.7.0
importlib_resources==6.5.2
iniconfig==2.0.0
instructor==1.13.0
jeepney==0.7.1
Jinja2==3.1.6
jiter==0.11.1
joblib==1.4.2
json5==0.12.1
json_repair==0.25.2
jsonref==1.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
keyring==23.5.0
kiwisolver==1.4.5
kubernetes==34.1.0
lance-namespace==0.0.20
lance-namespace-urllib3-client==0.0.20
lancedb==0.25.2
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
lxml==4.8.0
lz4==3.1.3+dfsg
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.3.6
markdown-it-py==4.0.0
MarkupSafe==2.0.1
matplotlib==3.8.0
mcp==1.21.0
mdurl==0.1.2
mmh3==5.2.0
monotonic==1.6
more-itertools==8.10.0
mpmath==0.0.0
multidict==6.7.0
netifaces==0.11.0
nltk==3.9.1
nodeenv==1.9.1
numpy==1.26.0
oauthlib==3.2.0
olefile==0.46
onnxruntime==1.23.2
openai==2.7.1
openpyxl==3.1.5
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
opentelemetry-exporter-otlp-proto-http==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
orjson==3.11.4
overrides==7.7.0
packaging==23.1
pandas==2.2.0
paramiko==2.9.3
pbr==6.0.0
pdfminer.six==20250506
pdfplumber==0.11.7
pexpect==4.8.0
pillow==12.0.0
pipenv==2022.7.24
platformdirs==2.5.2
plotly==5.20.0
pluggy==1.3.0
ply==3.11
portalocker==2.7.0
posthog==5.4.0
pre_commit==4.3.0
propcache==0.4.1
protobuf==6.33.0
ptyprocess==0.7.0
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.11
pybase64==1.4.2
pycairo==1.20.1
pycparser==2.23
pycups==2.0.1
pydantic==2.12.4
pydantic-settings==2.11.0
pydantic_core==2.41.5
pygame==2.5.2
Pygments==2.19.2
PyGObject==3.42.1
PyJWT==2.10.1
pylance==0.39.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pypdf==6.1.3
pypdfium2==5.0.0
PyPika==0.48.9
pyproject_hooks==1.2.0
PyQt5==5.15.6
PyQt5-sip==12.9.1
PyQtWebEngine==5.15.5
pyRFC3339==1.1
pyrsistent==0.18.1
PySocks==1.7.1
pytest==7.4.0
python-apt==2.4.0+ubuntu4
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-docx==1.2.0
python-dotenv==1.2.1
python-multipart==0.0.20
pythran==0.10.0
pytube==15.0.0
pytz==2022.1
pyxdg==0.27
PyYAML==6.0.3
referencing==0.37.0
regex==2024.11.6
reportlab==3.6.8
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.2.0
rpds-py==0.28.0
rsa==4.9.1
scipy==1.8.0
seaborn==0.13.2
SecretStorage==3.3.1
Send2Trash==1.8.1b0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.0
sse-starlette==3.0.3
starlette==0.50.0
stevedore==5.1.0
sympy==1.9
systemd-python==234
tenacity==8.2.3
textblob==0.19.0
tiktoken==0.12.0
tokenizers==0.22.1
tomli==2.3.0
tomli_w==1.2.0
torbrowser-launcher==0.3.3
tqdm==4.67.1
ty==0.0.1a25
typer==0.20.0
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2023.4
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufoLib2==0.13.1
ufw==0.36.1
unattended-upgrades==0.1
unicodedata2==14.0.0
urllib3==1.26.5
usb-creator==0.3.7
uv==0.9.7
uvicorn==0.38.0
uvloop==0.22.1
virtualenv==20.16.2
virtualenv-clone==0.5.7
virtualenvwrapper==6.0.0
wadllib==1.3.6
watchfiles==1.1.1
webencodings==0.5.1
websocket-client==1.9.0
websockets==15.0.1
xdg==5
xkit==0.0.0
yarl==1.22.0
youtube-transcript-api==1.2.3
zipp==3.23.0
